Fix AddEditResult snackbar sometimes not showing
Previously the result was retrieved as a plain variable when the UI state was updated, but the UI state only updated if the list of the notes changed.
Now it is retrieved as a flow and if either of the two changes, the UI state gets updated and the snackbar shown.