Add API response error checking to auth functions
The AuthApiService class simply makes the requests and returns with the deserialized response object. If the request fails, it throws different Ktor exceptions.

Those exceptions are caught in the AuthDataSource class, which calls the AuthApiService functions. It wraps the response in a Result<V, E> object, in which V is the type of the API return type and E is an ApiError type.

An instance of AuthDataSource gets injected into the sign in/register viewmodel, which gets the Result<V,E> object from the request and acts appropriately to the Result type.