Change how note-label relationship is stored
Firstly the note id field has been changed to UUID type similarly to the label id field.
Secondly the labels associated with a note are no longer stored as a json string, rather the association is stored in a separate table. This is a more elegant solution, but also it will be easier to work with in the future.