Add soft deleting notes and labels when immediate remote deletion fails
When deleting a note or a label, it first gets deleted from the server. If that is successful, it gets immediately deleted from the local database as well. But if the remote deletion fails, it first gets soft deleted, so when the next synchronization occurs, the server gets notified of the deletion. Soft deleted notes and labels get deleted from the local database after the first successful synchronization.